backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
PyJWT>=2.6.0
bcrypt>=4.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pillow>=9.0.0
